Linked columns must be same datatype.

+++ database.sql
@@ -53,7 +53,7 @@ CREATE TABLE IF NOT EXISTS `items` (
   /* Reduce VARCHAR after data entry. */
   `name` varchar(20) NOT NULL,
   `description` int(11) NOT NULL,
-  `type_id` int(11) NOT NULL,
+  `type_id` smallint(6) NOT NULL,
   PRIMARY KEY (`id`),
   FOREIGN KEY (`type_id`) REFERENCES item_types(`id`),
   UNIQUE KEY `name` (`name`)
